Return the created_at data of registered function api objects
