+++ services/web/server/tests/unit/with_dbs/04/functions_rpc/test_functions_controller_rest.py
@@ -21,6 +21,8 @@ from pytest_simcore.helpers.webserver_login import UserInfoDict
 from servicelib.aiohttp import status
 from simcore_service_webserver.db.models import UserRole
 
+pytest_simcore_core_services_selection = ["rabbit"]
+
 
 @pytest.fixture
 def mock_function() -> dict[str, Any]:
